arch/x84/kernel/timers: options to disable PIT, set tsc and apic freq.
Some platforms don't have anything like a PIT, and further
don't have any useful timebase with which to do calibration.
Further, the APIC frequency is known and time can be saved
by not calibrating it; it can be passed in on the command line.

This patch adds two command line options:

lapictimerfreq, which allows preconfiguration of the lapic timer
frequency from the kernel command line

pit, which allows the pit to be disabled via
pit=none

tscfreq is used to indicate the tsc frequency.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>