smm: we're kind of there
Here's the deal (and the problem)

coreboot installs the smm handler and, to simulate the UEFI
startup situationm, I have coreboot *not* locking down the
SMM space. I then install the relocater at 0x38000 and the
handler at 0xa0000.

When I run the SMI under coreboot, because SMI has run, we jump
right to 0xa0000, not 0x38000. That's actually OK.

There's more to it but I'll push this for now and we can
start debug stuff on real hardware soon.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>